:wrench: Rename envFile & move all declare
